Add library to local server
The library used for disabling autofill has been downloaded and stored ready for local use.

This is just in case there are issues with the remote location of the library, currently stored on what appears to be GitHub.

The ideal scenario is to use the remote one by default but use the locally stored example as a fallback, but this is something for the future.

Kudos to Terry Lin for the library.