Add build and test to pull_request
Experiment to see whether the pull_request is the right hook for Drone to build and unit test branches, as at the moment only a clone of the repo is performed.

Also added an LCOV reporter to the unit tests to generate coverage results, with the aim of SonarQube picking these up to be sent for coverage report stats.