Add more unit tests and small bits
Added unit tests to explore the possibility of controller based unit tests. Example used in post controller for departure page.

Tweaks to the test configurations to better reflect code coverage and to incorporate all tests in folder structures.

Kick starting the process of folding in i18n, in this case, to externalise string constants so that future labelling or translations can be done off a few text files rather than in code.

Some small SonarQube driven code refactors, like removing unused parameters, variables and splitting code blocks into smaller functions.

Begin laying the foundations for informing a user is unverified. This needs to be done in tandem with updated E2E tests.

Fixes and tweaks to the resend validation mail, as variables appear to be incorrect.

Refactored the generate hash for verification emails into the verification API class.