Update Drone to use commit SHAs
Instead of using current Drone build numbers, which means that a production deployment actually builds the entire codebase again and pushes another build onto Quay, artifacts are now using the commit SHA as tags. This makes it less human readable but potentially speeds up production deployments by skipping the build process.

The artifacts could be tagged with the build number as well, but this is commented out for now, so see whether there are any ill effects.