Tweak to Drone script
Added E2E DB secrets to the build and to environment variables for the E2E tests.

Grouped the unit test and SonarQube steps into a group for parallel execution. Up for debate whether unit test should be removed altogether (as it is run during the SonarQube step).

Some more SonarLint driven development.