Implement country autocomplete
For adding or editing person details, the person-details template should now leverage the GOVUK accessible autocomplete. It will fallback to a drop down selection should there be an issue with the Javascript.

The fallback should probably utilise more accessibility capability.

Essentially, country labels have the code appended on the end, so a user can type in country or code to reduce the options available.

The person-detail template is also has all its strings externalised so the process is now underway.

For hints that show an example date in the future, a little overengineering is in place to always display 2 years into the future, should be available across all templates if needed.

The next step is for unit and E2E tests for this functionality.