Implement port code converter
Taking the data from openflights.org for airports, a small piece of code to convert it into a form that could be used by sGAR, allowing potentially a user to search via IATA or ICAO codes for an airport.

It may be beneficial for data quality to prioritise ICAO codes in situations where IATA is used, this will be up for discussion.