Remove use strict declaration
SonarQube highlights it, although the use of ES6/ES2015 throughout the code implies the declaration is not really necessary, so removing it.

Attempted to refactor duplicates of simple GET controller.js code, into a nav.js utility class. This could be expanded to deal with more duplicate code ideally.

Removed the tabs includes within the common templates folder as it does not seem to be used (and something like the govuk components could be used in its stead).

Existing unit tests do not appear to be affected.