Add unit tests
Unit tests for the login post controller and all possible flows. Added an error thrown should the userApi return an unexpected message (it could have likely hung).

Externalised messages for the verification screen within the code. Will need to do so for the template.

Removed the generateHash function in verificationApi as it is the same as the one within create-token.

Some refactoring of the validator.js, which only highlights the redundant effort, as all the date validation should be handled with a library like moment.js.

Current issue with the unit tests is that with a deeper level of nested Promises, some of the sinon stubs and spies do not appear to detect calls. This needs further investigating, either at the tests or the code itself.