Update unit tests
Discovered that nested Promises can have chained then() callbacks, which fixes the assertions for the user login post tests.

Added some code to lower case the incoming username field when logging in, which should probably fix that issue as is. Will need to investigate and perhaps fix the underlying data in the DB.

Added back the country autocomplete for person details.

Also tweaked the airport list generation to munge two codes (IATA and ICAO) for an airport into a single row, with IATA codes taking preference. This approach would have implications for the existing data quality, so will need a question raised for business.