Create unit tests for upload gar function
Added unit tests for the uploadgar POST call, has meant splitting the POST functionality into a form that is actually consistent with the rest of the code, as it does not seem immediately obvious how to unit test them if within index.js code.

Unit test code has highlighted gaps in the functionality and possible improvements are noted within the unit tests.

Created a workbook-data file to hold XLSX compatible dummy data for those tests.