Experiment with routes
Possible to use something like /aircraft/next as a route and experiment with sessions. This still actually results in links though, and in that case probably better to leave it with query parameters.

Have disabled the referer check in the usercheck middleware, which allows a user to change page number. Whether this is the final approach going forward, not sure.