Add validations for dates for Excel upload
Manifests of crew and passengers had only a non-empty check when uploading, but did not handle non date inputs, despite the Excel spreadsheet allowing them. This resulted in completely missing manifests in the upload.

This should be resolved, via the use of an additional validator to ensure the incoming value is valid via moment.js. This may be tweaked in the future with other merges.