Change from query parameter to session variable
This has resulted in no need for passing the query parameter across pages at the expense of being able to transition to pages by url. It has also meant that only parts of the express-paginate library was useful, so has been dropped.

Pagination module is now in the common/utils folder. Will need unit testing.