EGAR-764 or DSUP-3521 add message for empty tables
For tables within the forms, if there are no records, a message is displayed instead of an empty table. Current approach is to display the message within the table, with the exception of the people screen, which demonstrates how the message would look without the table.

Will be up to business to determine which looks most suitable.

Cleaned up some externalised strings, some column widths, but also need to revisit the summary page as the two tables have maximum width.