Externalise pagination markup
Currently first attempt at making the pagination resuable, via the use of a macro. It needs to take in the object returned from the pagination.js build function, in addition to the i18n handler, which seems unnecessary, so could do with some refinement.

However, it does appear to work, so is now ready for an attempt for the manage people screen.

Unit tests also added for the pagination module, which only requires some edge cases to be added.