Fix for duplicate aircraft adding
Possibly fix which absorbs the error returned from the back end should a user add an aircraft that another user has added. This prevents the app from hanging, but still needs a fix on the back end.

Refactored the header so that there is not so much duplication. Use of ids for easier handling in E2E tests, and externalised labels.