Departure port now uses YYYY
Instead of ZZZZ causing a clash with the arrival port, departure port now uses YYYY instead.

This is now informed to the user, and will change to YYYY automatically even if user enters ZZZZ.

Whether this actually works as expected with CBP is still an unknown quantity, hence the lack of extensive testing.

E2E tests are not changed, mainly due to CBP, but also because future stories will improve the UX around this negating the need for a user to enter ZZZZ or YYYY manually.

Also corrected lattitude in the function name.