Refactor startup
Initialise DB code was previously in the start.js, and also meant it was run twice due to be being part of the two worker processes. It is now within the server.js and hopefully performed once.

Refactored unit tests so the unhandledRejection handler is within a global.test file and also removed after tests. This reduces some duplication of code.

Added unit tests for edge logic in Manifest.class.js and moved into the appropriate folder.