Possible approach to access control
Should the usercheck removal of the referer header be the only solution to Edge browser issues, then some sort of access control is required for the organisation screens.

This is an attempt to do so, via a pair of middleware determined by the presence of the organisation id in a user session.

This is simplistic, but is also the only evidence of user role distinction in the app, so is most likely sufficient.