Possible screen reader tweaks for login and registration
Removed the role button for the sign in, as it really is just a link rather a button to submit something. Leaving the button styling as it is the positive action for their respective screens.

Added aria-describedby attributes to hopefully aid input text inputs.

Refactored some variables, like 'Username' to 'username'.

Removed the unobtrusive validation attributes in the registration screen, as it is not used.

Externalised the strings for the login and registration screens.