Add error prefixes to titles for login and flight pages
"Error: " prefix on titles so a screen reader reads it out as soon as the error messages appear for a page.

Externalised the prefix (which is hidden) for normal in line messages. This has meant a change in the call to the macro that builds it to include a handle for the i18n library.

Added error hints to the date and time fields.

Renamed mfa-code to mfaCode so that bracket notation is not used in the post controller for MFA verification page (cosmetic rather than anything else).

Adding aria-describedby attributes for fields in the registration page.