Experiment with skipping login if still logged in
Session lasts a day, and when a user goes to the site, they are still expected to log in again, even if they might have say typed into the address bar. There is now a check to see if the session exists and goes to the dashboard from the welcome page straight away.

Session TTL will need to be checked, currently appears to be based off of the cookie time of 1 hour, but needs double checking.

Also some indentation cleanup of the aircraft pages ready for reviewing screen reader issues.