Externalise strings in welcome page
Welcome page has lots of text, so previous text is commented out to be ignored by Nunjucks, but kept in for now as the page could be considered difficult to understand. Could be removed in future.

Link to the website page where the GAR form can be downloaded now externalised, which will be very useful should that link change in the future.