Experiment with pathing
Departure screen has three buttons, two using the secondary formatting, where the exit path now goes to the GAR Summary View instead of the home page. Secondary formatting copied over from GOV UK stylesheets.

The back button on said GAR Summary now goes to the Dashboard Page.

This means 'Save and exit' functions will need to redirect with a 307 code to pass itself as a POST to the garfile/view which relies on a POST to display its contents, something which may need revising in future (change to GET and using sessions possibly).

Change to the GAR Summary controller to take the GAR ID from the cookie should it not be in the body. Again, this may benefit from a change to a GET controller taking the property elsewhere.

Session Postgres settings explicitly set to 1 hour to match the cookie maxAge.

Removed the unsubmitted/submitted templates for GAR Summary view page as they are virtually duplicates.

Added a "opens in new window" to a link for accessibility purposes.

Unit tests updated.