Tweaks to ease E2E test scripting
Adding ids in a few places to allow for Selenium to find page elements easier. Removed the garId as an input because the GAR Summary post controller does the same thing, and is a security risk.

On the cancel GAR page, exiting without cancelling will return to the summary rather than the dashboard.

Externalised strings for Cancel GAR page.