WAVE driven development
Empty table messages are a <caption> and a macro so is reusable.

Conditional inclusion of aria-describedby for errors shown for the departure and arrival screen. This has involved the creation of a nunjucks filter to check whether a specific error is present in order to render the appropriate attribute. This would otherwise be considered an error by WAVE.

Externalised strings for the organisation home page.

Externalised strings for the invite user page.

Invite users and also now the register screen take the input values and return them on error so they remain populated. For the register user screen, the cookie is not used to prepopulate.

Moved the resend token link to after the submit button.

Attempts to clean up the manifest person further details markup for WAVE, but still picks it up as an alert.

Reverted the link at the top to go home rather the GAR dashboard. WAVE highlighted this as two redundant adjacent links, and with the tweaks to signing up not requiring another log in if a session exists, the original bug is not applicable. There appears to be no distinction between pre-user and user headers, so these need refactoring in the future.

Added and tweaked unit tests.