Update Responsible Person page for accessibility
Folded in tweaks from other pages to the responsible person page.
Addressed ambiguity in telephone validation message and folded in error message improvements.
Corrected the redirect for arrival and responsible person page so that it goes to the GAR Summary screen.
When creating a new GAR, session is explicitly saved to prevent error message appearing.
Externalised strings for responsible person page.
Markup tweaks for the GAR dashboard, some erroneous whitespaces removed.
User registration submission would set the UserDbId into the cookie, which it should not do because the user is not verified at that stage.
Upon verifying an MFA token, there is now an explicit session save call to ensure the cookie has the new login info stored before redirecting to home (and potentially back out of the app).
Removed get and set inviteOrgId, as these are not called anywhere in the code base.
Added the :focus CSS styling for secondary buttons, as focus previously made it the same green background colour as the default button.
Updated unit tests.