Add accessibility tweaks to the people pages
Beginning the a11y work for the people pages. One initial action is to rename some of the variables for consistency. This has meant some updating of unit tests as a result.

Have added a few of the appropriate aria-describedby labels, but will need another revision as not sure it is comprehensive.

Tweaks to the submission buttons for adding a new person, as a table to arrange things is out of fashion.