Add customs page validation for GAR review
Checks for null values for the customs page fields to prevent submission if they are absent.
Externalised strings for the rest of the review page.
Optimised paths for the review page.
Added ability to enter GAR review page from the Summary as an alternative way to submit a GAR.
Updated unit tests, will require E2E updates.