Add logic to display error message for unregistered user
Instead of an email and going to the MFA token screen, an error message is displayed instead. This message is externalised.
Added extra logic to ensure the "Error: " prefix is added to the title.
Some removal of duplicate code, in a few of the index.js files and a block in the garfile/craft GET controller class.
Removed duplicate macros import a few more nunjuck template files.
Updated unit tests.