Implement possible approach to set nav links as active
Using just nunjucks, set a variable called "active" to be picked up by the nav link and used to check against the href. This allows for groups of pages to use the same variable and highlight the same nav link. This will mean the variable will need setting throughout the templates.

No unit tests as this is nunjucks only, so E2E tests are expected to ensure this is as expected.