Add radio buttons for departure screen
Rudimentary implementation of the radio buttons to select a choice of port or co-ordinates. Presented as a question and conditionally displays the port or co-ordinate inputs.

This still requires validation tweaks (including a new one regarding ensuring a selection is made), and accessibility testing.

There is also the requirement of some additional logic to handle the co-ordinate input, which will also set the port code to be 'ZZZZ' (or something else, still waiting on requirements).