Updates to GAR API Promise handling
Rather than log out aspects of the error, reject with the error to be picked up by the caller. This should hopefully improve the logging output, but no guarantee. Regardless, it removes the redundant catch block for each method and appears more in line with expected Promise handling.

Also some screen reader tweaks for the radio buttons on departure screen.