Fix csrf issue with aircraft pagination
WIthin the GAR aircraft page, csrf token issues were occurring because of the clash with the disableAutoFill library, which scrambles form names, and the csrf one in this case. This has been disabled for this page to allow the pagination to work.

Also reduced log level for a validation error in the login screen.