Add experimental glossary of terms page
Page is called terms and added to the app with appropriate routes. Includes some example terms and how they could be linked in the departure page.

The markup to create a link to the glossary and support screen readers is now a macro, so this could be used throughout the app.

If possible, a custom filter would help in replacing abbreviations with markup, as it currently lacks a little elegance although it does work.

Base template updated to include the terms page at the bottom. The footer will need externalising.